+++ changelog.md
@@ -25,6 +25,7 @@ Date format is `MM/DD/YYYY`.
 - delete() method to send a DELETE request
 - declineJoinRequests() method to batch decline join requests
 - getJoinRequests() method to get a list of join requests
+- acceptJoinRequests() method to batch accept join requests
 
 ### Fixed